(BSR)[API] chore: move dms out of scripts into core
