+++ api/src/pcapi/scripts/subscription/dms/__init__.py
@@ -1,2 +1,1 @@
 from .handle_inactive_dms_applications import handle_inactive_dms_applications
-from .import_dms_applications import import_all_updated_dms_applications
